Isolate sticky lines processor for unit testing
